+++ docker/docker-compose.prod.yml
@@ -8,8 +8,6 @@ services:
       context: .
     restart: always
     ports: [80:80]
-    # JDK 8 doesn't know how to manage its memory when it's in a container. JDK 9 fixed this, but Kotlin only supports
-    # JDK 8. Hence, we have to use the relevant compiler options.
     command: java
       -server
       -XX:+UnlockExperimentalVMOptions
