+++ docs/auth_admin_panel.md
@@ -1,3 +1,3 @@
-# Auth
+# Auth Admin Panel
 
 The auth system's admin panel runs on http://localhost:8080/auth/admin. The username is `admin`, and the password is the value of `KEYCLOAK_PASSWORD` in `.env`.
