+++ docs/developing.md
@@ -1,8 +1,5 @@
 # Developing
 
-- The auth system disallows two users from having the same registered email address.
-- A "private chat" is a chat between two people which cannot be converted into a group chat.
-
 ## Flow
 
 Here's how a standard project iteration looks like.
@@ -12,12 +9,12 @@ Here's how a standard project iteration looks like.
 1. Create any required [models](../src/main/kotlin/Models.kt).
 1. If you're updating the DB, keep in mind that you might have to wipe it when the account is deleted.
 1. Write tests (i.e., TDD). If you're writing tests for an HTTP API endpoint, perform the following sub-steps.
-    1. Create a file in [`src/test/kotlin/routes`](../src/test/kotlin/routes) named using the format `<TAG>Test.kt`, where `<TAG>` is the feature's tag in the OpenAPI spec. For example, the JWT feature's `/jwt-request` and `/jwt-refresh` endpoints are tagged `jwt` in the OpenAPI spec, and are therefore have their tests in [`src/test/kotlin/routes/JwtTest.kt`](../src/test/kotlin/routes/JwtTest.kt).
-    1. Create functions for each endpoint in the newly created file, each of which return a `io.ktor.server.testing.TestApplicationResponse`. The functions should be named the operation ID used by the endpoint in the OpenAPI spec.
-    1. Create classes for the endpoints using the format `<HTTP_VERB><ENDPOINT>Test`. For example, if the endpoint `/chat-message` accepts the POST and PATCH verbs, then two classes named `PostChatMessageTest` and `PatchChatMessageTest` would be created.
+    1. Create a file in [`src/test/kotlin/routes`](../src/test/kotlin/routes) named using the format `<URL>Test.kt`. For example, the endpoint `/jwt-request` has its tests in [`src/test/kotlin/routes/JwtRequestTest.kt`](../src/test/kotlin/routes/JwtRequestTest.kt).
+    1. Create a function in the newly created file which returns a `io.ktor.server.testing.TestApplicationResponse`. The function should be named the operation ID used by the endpoint in the OpenAPI spec.
+    1. Create a class using the format `<HTTP_VERB><URL>Test`. For example, if the endpoint `/chat-message` accepts the POST verb, then the class `PostChatMessageTest` would be created.
 1. Implement the feature.
 
-    If the feature is an HTTP API endpoint, implement it in files created mirroring the test source set's structure. For example, the JWT feature's tests are in [`src/test/kotlin/routes/JWTTest.kt`](../src/test/kotlin/routes/JwtTest.kt), and therefore its implementation is in [`src/main/kotlin/routes/JWT.kt`](../src/main/kotlin/routes/Jwt.kt).
+    If the feature is an HTTP API endpoint, then name the function the same as its operation ID in the OpenAPI spec. If the endpoint accepts multiple HTTP verbs, create a separate routing function named using the format `route<URL>`. See [`src/main/kotlin/routes/Account.kt`](../src/main/kotlin/routes/Account.kt) for an example.
 1. If the feature was from the [spec](spec.md), mark it as completed.
 1. If you have updated the server's functionality, or the OpenAPI spec, follow these sub-steps to create a new release.
     1. Update the version in the [OpenAPI spec](openapi.yaml) and the [build file](../build.gradle.kts). Even if the server is still backwards compatible, you must bump the major version if you've renamed an entity in the OpenAPI spec (e.g., a key under `schemas/components/`). This is because [OpenAPI Generator](https://openapi-generator.tech/) uses the names of keys when creating client SDKs, which would have otherwise become backwards-incompatible.
@@ -26,7 +23,7 @@ Here's how a standard project iteration looks like.
 
 ## Server
 
-### Abstraction Layers
+### Notes
 
 - Always use the [`src/main/kotlin/Auth.kt`](../src/main/kotlin/Auth.kt) and [`src/test/kotlin/Auth.kt`](../src/test/kotlin/Auth.kt) abstraction layers instead of directly using Keycloak's API.
 - Always use the [`src/main/kotlin/db/`](../src/main/kotlin/db) and [`src/test/kotlin/db/`](../src/test/kotlin/db) abstraction layers instead of directly operating on the DB.
@@ -55,7 +52,7 @@ Here's how a standard project iteration looks like.
         --project-directory . \
         up --scale chat=0 -d
     ```
-1. Enter into the shell. The connection will be refused continuously for approximately one minute until the required services are up and running. You needn't have a valid email configuration in `.env` (e.g., `KEYCLOAK_SMTP_HOST`, `KEYCLOAK_SMTP_PASSWORD`), because the testing environment disables the sending of emails.
+1. Enter into the shell. You needn't have a valid email configuration in `.env` (e.g., `KEYCLOAK_SMTP_HOST`, `KEYCLOAK_SMTP_PASSWORD`), because the testing environment disables the sending of emails.
     ```
     docker-compose \
         -f docker/docker-compose.yml \
@@ -65,11 +62,14 @@ Here's how a standard project iteration looks like.
         run --rm chat bash
     ```
 1. Test by running `gradle test` whenever you want. Build reports save to `build/reports/tests/test/`.
-1. Run `exit` once you're done.
 
 ### [Production](production.md)
 
-## [Auth](auth.md)
+## Auth
+
+- The auth system disallows two users from having the same registered email address.
+- The auth system strips leading and trailing whitespace for usernames, first names, and last names.
+- There is an [admin panel](auth_admin_panel.md).
 
 ## [Spec](spec.md)
