+++ src/main/kotlin/db/GroupChats.kt
@@ -3,17 +3,20 @@ package com.neelkamath.omniChat.db
 import com.neelkamath.omniChat.Auth
 import com.neelkamath.omniChat.GroupChat
 import com.neelkamath.omniChat.GroupChatUpdate
+import com.neelkamath.omniChat.db.GroupChats.adminUserId
 import org.jetbrains.exposed.dao.id.IntIdTable
 import org.jetbrains.exposed.sql.insertAndGetId
 import org.jetbrains.exposed.sql.select
+import org.jetbrains.exposed.sql.selectAll
 import org.jetbrains.exposed.sql.statements.UpdateStatement
 import org.jetbrains.exposed.sql.update
 
 data class GroupChatWithId(val id: Int, val chat: GroupChat)
 
+/** The [GroupChatUsers] table contains the participants, including the [adminUserId], of a particular chat. */
 object GroupChats : IntIdTable() {
     override val tableName get() = "group_chats"
-    val adminUserId = varchar("admin_user_id", Auth.userIdLength)
+    private val adminUserId = varchar("admin_user_id", Auth.userIdLength)
     const val maxTitleLength = 70
     val title = varchar("title", maxTitleLength)
     const val maxDescriptionLength = 1000
@@ -35,7 +38,7 @@ object GroupChats : IntIdTable() {
             it[title] = chat.title
             it[description] = chat.description
         }.value
-        GroupChatUsers.create(groupId, chat.userIdList)
+        GroupChatUsers.create(groupId, chat.userIdList + adminUserId)
         groupId
     }
 
@@ -53,5 +56,17 @@ object GroupChats : IntIdTable() {
         }
         update.newUserIdList?.let { GroupChatUsers.addUsers(update.chatId, it) }
         update.removedUserIdList?.let { GroupChatUsers.removeUsers(update.chatId, it) }
+    }
+
+    /**
+     * Searches the chats the [userId]'s is in.
+     *
+     * Returns the chat ID list by searching for the [query] in every chat's title case-insensitively.
+     */
+    fun search(userId: String, query: String): List<Int> = Db.transact {
+        selectAll()
+            .filter { it[id].value in GroupChatUsers.getChatIdList(userId) }
+            .filter { query.toLowerCase() in it[title].toLowerCase() }
+            .map { it[id].value }
     }
 }
