+++ src/main/kotlin/routes/JwtRequest.kt
@@ -1,11 +1,9 @@
 package com.neelkamath.omniChat.routes
 
-import com.auth0.jwt.JWT
 import com.neelkamath.omniChat.*
 import io.ktor.application.call
 import io.ktor.http.HttpStatusCode
 import io.ktor.request.receive
-import io.ktor.request.receiveParameters
 import io.ktor.response.respond
 import io.ktor.routing.Routing
 import io.ktor.routing.post
@@ -28,13 +26,5 @@ fun Routing.requestJwt() {
                 }
             }
         }
-    }
-}
-
-fun Routing.refreshJwt() {
-    post("jwt-refresh") {
-        val token = call.receiveParameters()["refresh_token"]!!
-        val userId = JWT.decode(token).subject
-        call.respond(Jwt.buildAuthToken(userId, Auth.refreshToken(token)))
     }
 }
