+++ src/test/kotlin/routes/AccountTest.kt
@@ -6,7 +6,10 @@ import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.booleans.shouldBeFalse
 import io.kotest.matchers.shouldBe
 import io.ktor.application.Application
-import io.ktor.http.*
+import io.ktor.http.ContentType
+import io.ktor.http.HttpHeaders
+import io.ktor.http.HttpMethod
+import io.ktor.http.HttpStatusCode
 import io.ktor.server.testing.TestApplicationResponse
 import io.ktor.server.testing.handleRequest
 import io.ktor.server.testing.setBody
@@ -34,16 +37,6 @@ fun readAccount(jwt: String): TestApplicationResponse = withTestApplication(Appl
 
 fun deleteAccount(jwt: String): TestApplicationResponse = withTestApplication(Application::main) {
     handleRequest(HttpMethod.Delete, "account") { addHeader(HttpHeaders.Authorization, "Bearer $jwt") }
-}.response
-
-fun verifyEmail(email: String): TestApplicationResponse = withTestApplication(Application::main) {
-    val parameters = Parameters.build { append("email", email) }.formUrlEncode()
-    handleRequest(HttpMethod.Get, "email-verification?$parameters")
-}.response
-
-fun resetPassword(email: String): TestApplicationResponse = withTestApplication(Application::main) {
-    val parameters = Parameters.build { append("email", email) }.formUrlEncode()
-    handleRequest(HttpMethod.Get, "password-reset?$parameters")
 }.response
 
 class GetAccountTest : StringSpec({
@@ -170,33 +163,5 @@ class DeleteAccountTest : StringSpec({
         Auth.verifyEmail(login.username)
         deleteAccount(getJwt(login)).status() shouldBe HttpStatusCode.NoContent
         Auth.usernameExists(login.username).shouldBeFalse()
-    }
-})
-
-class GetEmailVerificationTest : StringSpec({
-    listener(AppListener())
-
-    "Sending a verification email should respond with an HTTP status code of 204" {
-        val email = "username@example.com"
-        createAccount(NewAccount("username", "password", email))
-        verifyEmail(email).status() shouldBe HttpStatusCode.NoContent
-    }
-
-    "Sending a verification email to an unregistered address should respond with an HTTP status code of 400" {
-        verifyEmail("username@example.com").status() shouldBe HttpStatusCode.BadRequest
-    }
-})
-
-class GetPasswordResetTest : StringSpec({
-    listener(AppListener())
-
-    "Requesting a password reset should respond with an HTTP status code of 204" {
-        val email = "username@example.com"
-        createAccount(NewAccount("username", "password", email))
-        resetPassword(email).status() shouldBe HttpStatusCode.NoContent
-    }
-
-    "Requesting a password reset for an unregistered address should respond with an HTTP status code of 400" {
-        resetPassword("username@example.com").status() shouldBe HttpStatusCode.BadRequest
     }
 })
