+++ src/test/kotlin/routes/JwtRequestTest.kt
@@ -5,7 +5,10 @@ import com.neelkamath.omniChat.test.verifyEmail
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
 import io.ktor.application.Application
-import io.ktor.http.*
+import io.ktor.http.ContentType
+import io.ktor.http.HttpHeaders
+import io.ktor.http.HttpMethod
+import io.ktor.http.HttpStatusCode
 import io.ktor.server.testing.TestApplicationResponse
 import io.ktor.server.testing.handleRequest
 import io.ktor.server.testing.setBody
@@ -15,13 +18,6 @@ fun requestJwt(login: Login): TestApplicationResponse = withTestApplication(Appl
     handleRequest(HttpMethod.Post, "jwt-request") {
         addHeader(HttpHeaders.ContentType, ContentType.Application.Json.toString())
         setBody(gson.toJson(login))
-    }
-}.response
-
-fun refreshJwt(refreshToken: String): TestApplicationResponse = withTestApplication(Application::main) {
-    handleRequest(HttpMethod.Post, "jwt-refresh") {
-        addHeader(HttpHeaders.ContentType, ContentType.Application.FormUrlEncoded.toString())
-        setBody(listOf("refresh_token" to refreshToken).formUrlEncode())
     }
 }.response
 
@@ -61,19 +57,5 @@ class PostJwtRequestTest : StringSpec({
         response.status() shouldBe HttpStatusCode.BadRequest
         val body = gson.fromJson(response.content, InvalidAccount::class.java)
         body shouldBe InvalidAccount(InvalidAccountReason.INCORRECT_PASSWORD)
-    }
-})
-
-class PostJwtRefreshTest : StringSpec({
-    listener(AppListener())
-
-    "A refresh token should issue a new token set" {
-        val login = Login("username", "password")
-        createAccount(NewAccount(login.username, login.password, "username@example.com"))
-        Auth.verifyEmail(login.username)
-        val token = gson.fromJson(requestJwt(login).content, AuthToken::class.java).refreshToken
-        val response = refreshJwt(token)
-        response.status() shouldBe HttpStatusCode.OK
-        gson.fromJson(response.content, AuthToken::class.java) // Successfully parsing it verifies the response body.
     }
 })
