+++ src/test/kotlin/routes/UserSearchTest.kt
@@ -5,7 +5,6 @@ import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
 import io.ktor.application.Application
 import io.ktor.http.HttpMethod
-import io.ktor.http.HttpStatusCode
 import io.ktor.http.Parameters
 import io.ktor.http.formUrlEncode
 import io.ktor.server.testing.TestApplicationResponse
@@ -22,11 +21,6 @@ fun searchUsers(query: UserSearchQuery): TestApplicationResponse = withTestAppli
         }.formUrlEncode()
     }
     handleRequest(HttpMethod.Get, "user-search?$parameters")
-}.response
-
-fun readUser(userId: String): TestApplicationResponse = withTestApplication(Application::main) {
-    val parameters = Parameters.build { append("user_id", userId) }.formUrlEncode()
-    handleRequest(HttpMethod.Get, "user?$parameters")
 }.response
 
 class GetUserSearchTest : StringSpec({
@@ -56,26 +50,5 @@ class GetUserSearchTest : StringSpec({
         nameResults shouldBe UserIdList(setOf(userIdList.elementAt(1), userIdList.elementAt(3)))
         val combinedResults = search(UserSearchQuery(username = "john", firstName = "john"))
         combinedResults shouldBe UserIdList(setOf(userIdList.elementAt(1)))
-    }
-
-    "Sending at least one filter should be made mandatory" {
-        searchUsers(UserSearchQuery()).status() shouldBe HttpStatusCode.BadRequest
-    }
-})
-
-class GetUserTest : StringSpec({
-    listener(AppListener())
-
-    "A user's info should be retrieved" {
-        val account = NewAccount("username", "password", "username@example.com", "first name")
-        createAccount(account)
-        val response = readUser(Auth.findUserByUsername(account.username).id)
-        response.status() shouldBe HttpStatusCode.OK
-        val body = gson.fromJson(response.content, User::class.java)
-        body shouldBe User(account.username, account.email, account.firstName, account.lastName)
-    }
-
-    "Requesting the details of a nonexistent user should respond with an HTTP status code of 400" {
-        readUser("a nonexistent user ID").status() shouldBe HttpStatusCode.BadRequest
     }
 })
